script/requirements.txt
azure-ai-textanalytics==5.3.0
azure-common==1.1.28
azure-core==1.30.2
certifi==2024.7.4
charset-normalizer==3.3.2
coverage==7.6.1
idna==3.7
isodate==0.6.1
praw==7.7.1
prawcore==2.4.0
requests==2.32.3
setuptools-scm==6.3.2
six==1.16.0
typing_extensions==4.12.2
update-checker==0.18.0
urllib3==2.2.2
websocket-client==1.8.0
parameterized==0.9.0
azure.communication.email==1.0.0